admin-type(list,edit,delete,add)-ok admin-category(list,edit,delete,add)-on going

+++ src/main/java/com/my_project/niit_final_project/entities/Category.java
@@ -18,7 +18,7 @@ public class Category {
     String name;
 
     @Column(name = "status")
-    Long status;
+    Integer status;
 
     @Column(name = "product_type_id")
     Long typeId;
